requirements.txt
asgiref==3.6.0
Django==4.1.7
django-phone-field==1.8.1
Pillow==9.4.0
sqlparse==0.4.3
